feat: 화면 이동 리스너 분리
