+++ app/src/main/java/com/woowa/banchan/ui/OnOrderClickListener.kt
@@ -3,5 +3,4 @@ package com.woowa.banchan.ui
 interface OnOrderClickListener {
 
     fun navigateToOrder()
-    fun navigateToOrderDetail()
 }
