+++ app/src/main/java/com/woowa/banchan/ui/OnOrderDetailClickListener.kt
@@ -1,0 +1,6 @@
+package com.woowa.banchan.ui
+
+interface OnOrderDetailClickListener {
+
+    fun navigateToOrderDetail()
+}
