+++ app/src/main/java/com/woowa/banchan/ui/main/MainFragment.kt
@@ -83,16 +83,6 @@ class MainFragment : Fragment(), OnDetailClickListener, OnCartClickListener, OnO
             .commit()
     }
 
-
-    override fun navigateToOrderDetail() {
-        parentFragmentManager.popBackStack("Main", FragmentManager.POP_BACK_STACK_INCLUSIVE)
-        parentFragmentManager.beginTransaction()
-            .setCustomAnimations(R.anim.slide_in, 0, 0, R.anim.slide_out)
-            .addToBackStack("Main")
-            .add(R.id.fcv_main, OrderDetailFragment())
-            .commit()
-    }
-
     override fun onDestroyView() {
         _binding = null
         super.onDestroyView()
