+++ app/src/main/java/com/woowa/banchan/ui/order/OrderFragment.kt
@@ -5,14 +5,18 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
+import com.woowa.banchan.R
 import com.woowa.banchan.databinding.FragmentOrderBinding
 import com.woowa.banchan.ui.OnBackClickListener
+import com.woowa.banchan.ui.OnOrderDetailClickListener
 import com.woowa.banchan.ui.extensions.repeatOnLifecycle
+import com.woowa.banchan.ui.orderdetail.OrderDetailFragment
 import com.woowa.banchan.ui.orderdetail.OrderDetailViewModel
 
-class OrderFragment: Fragment() {
+class OrderFragment: Fragment(), OnOrderDetailClickListener {
 
     private var _binding: FragmentOrderBinding? = null
     private val binding: FragmentOrderBinding get() = requireNotNull(_binding)
@@ -22,7 +26,7 @@ class OrderFragment: Fragment() {
     private val orderListAdapter by lazy {
         OrderListAdapter(
             onClickItem = {
-
+                navigateToOrderDetail()
             }
         )
     }
@@ -61,6 +65,15 @@ class OrderFragment: Fragment() {
                 }
             }
         }
+    }
+
+    override fun navigateToOrderDetail() {
+        parentFragmentManager.popBackStack("Order", FragmentManager.POP_BACK_STACK_INCLUSIVE)
+        parentFragmentManager.beginTransaction()
+            .setCustomAnimations(R.anim.slide_in, 0, 0, R.anim.slide_out)
+            .addToBackStack("Order")
+            .add(R.id.fcv_main, OrderDetailFragment())
+            .commit()
     }
 
     override fun onDestroyView() {
